+++ com.microej.example.iot.zwave/src/main/java/com/microej/example/iot/zwave/ControllerManager.java
@@ -1,8 +1,9 @@
 /*
  * Java
  *
- * Copyright 2016 IS2T. All rights reserved.
- * Use of this source code is subject to license terms.
+ * Copyright 2016-2019 MicroEJ Corp. All rights reserved.
+ * For demonstration purpose only.
+ * MicroEJ Corp. PROPRIETARY. Use is subject to license terms.
  */
 package com.microej.example.iot.zwave;
 
@@ -65,8 +66,8 @@ public class ControllerManager implements RegistrationListener<Controller> {
 	private void addHandlers(Controller controller) {
 		if (controller.getListener() instanceof EventControllerListener) {
 			EventControllerListener eventControllerListener = (EventControllerListener) controller.getListener();
-			eventControllerListener.addEventHandler("ej.basedriver.SwitchWithReturnState", switchStateManager);
-			eventControllerListener.addEventHandler("ej.basedriver.DryContact", dryContactManager);
+			eventControllerListener.addEventHandler("ej.basedriver.SwitchWithReturnState", switchStateManager); //$NON-NLS-1$
+			eventControllerListener.addEventHandler("ej.basedriver.DryContact", dryContactManager); //$NON-NLS-1$
 		}
 	}