+++ com.microej.example.iot.zwave/CHANGELOG.md
@@ -1,12 +1,34 @@
-## 1.0.1 (2017-08-07)
-Buf Fixes:
+# Changelog
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## 1.1.0 - 2019-09-09
+
+### Changed
+
+  - Update to MMM.
+  
+## 1.0.1 - 2017-08-07
+
+### Fixed
+
   - Change the DriverBaseZwave version.
 
-## 1.0.0 (2016-12-29)
-Features:
+## 1.0.0 - 2016-12-29
+
+### Added
+
   - Change organization.
   
-## 0.1.0 (2016-10-03)
-Features:
+## 0.1.0 - 2016-10-03
+
+### Added
+
   - Initial revision.
 
+---  
+_Markdown_   
+_Copyright 2019 MicroEJ Corp. All rights reserved._   
+_For demonstration purpose only._   
+_MicroEJ Corp. PROPRIETARY. Use is subject to license terms._